fix: unify sidebar switch
